requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.3
anyio==3.6.2
apturl==0.5.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
cmake==3.26.4
colorama==0.4.4
command-not-found==0.3
crypto==1.4.1
cryptography==3.4.8
cupshelpers==1.0
datasets==2.13.0
dbus-python==1.2.18
decouple==0.0.7
defer==1.0.6
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
docker-composer==2.17.4
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
email-validator==2.0.0.post1
evaluate==0.4.0
fastapi==0.95.1
fastapi-sessions==0.3.2
fasteners==0.14.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.2
gdown==4.7.1
greenlet==2.0.2
gyp==0.1
h11==0.14.0
httpcore==0.17.0
httplib2==0.20.2
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jose==1.0.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
lockfile==0.12.2
loguru==0.7.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
Naked==0.1.32
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
orjson==3.8.10
packaging==23.1
pandas==2.0.2
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.6.18
platformdirs==3.8.0
protobuf==3.12.4
ptyprocess==0.7.0
pyarrow==12.0.1
pycairo==1.20.1
pycryptodome==3.18.0
pycups==2.0.1
pydantic==1.10.7
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
regex==2023.6.3
reportlab==3.6.8
requests==2.25.1
responses==0.18.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
starlette==0.26.1
sympy==1.12
systemd-python==234
texttable==1.6.4
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wadllib==1.3.6
watchfiles==0.19.0
websocket-client==1.2.3
websockets==11.0.2
xdg==5
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.2
zipp==1.0.0
